backend/requirements.txt
absl-py==2.1.0
accelerate==1.0.0
aext-assistant
aext-assistant-server
aext-core
aext-core-server
aext-panels
aext-panels-server
aext-share-notebook
aext-share-notebook-server
aext-shared
aiohttp
aiosignal
anaconda-anon-usage
anaconda-catalogs
anaconda-client
anaconda-cloud-auth
anaconda-navigator
anaconda-project
annotated-doc==0.0.4
anyio
appdirs==1.4.4
appnope
archspec
argon2-cffi
argon2-cffi-bindings
asttokens
astunparse==1.6.3
async-lru
attrs
Babel
backports.functools-lru-cache
backports.tempfile
backports.weakref==1.0.post1
beautifulsoup4
bleach
blis==0.7.11
blobfile==3.1.0
bokeh
boltons
Bottleneck
Brotli
cachetools==5.4.0
catalogue==2.0.10
certifi
cffi
chardet
charset-normalizer
click
cloudpathlib==0.18.1
cloudpickle
clyent==1.2.2
comm
conda
conda-build
conda-content-trust
conda-libmamba-solver
conda-pack
conda-package-handling
conda-repo-cli==1.0.75
conda-token
conda-verify==3.4.2
conda_index
conda_package_streaming
confection==0.1.5
contourpy
coverage==7.11.0
cryptography
cycler==0.12.1
cymem==2.0.8
cytoolz
dask
dataretrieval==1.0.9
datasets==3.0.0
debugpy
decorator
defusedxml==0.6.0
dill==0.3.8
distributed
distro
dmglib
docker-pycreds==0.4.0
entrypoints
evaluate==0.4.3
executing
fastapi==0.121.1
fastjsonschema
filelock
flatbuffers==24.3.25
fonttools==4.53.0
frozendict
frozenlist
fsspec
future
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
h11
h5py==3.11.0
HeapDict
httpcore
httplib2==0.22.0
httpx
huggingface-hub==0.24.7
idna
importlib-metadata
iniconfig==2.3.0
intake
ipykernel
ipython
ipywidgets
jaraco.classes
jedi
Jinja2
joblib==1.4.2
json5
jsonpatch
jsonpointer==2.1
jsonschema
jsonschema-specifications
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab_server
jupyterlab_widgets
keras==3.4.1
keras-models==0.0.7
keyring
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
libarchive-c
libclang==18.1.1
libmambapy
lmdb
locket
lxml
lz4
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe
matplotlib==3.9.0
matplotlib-inline
mdurl==0.1.2
menuinst
mistune
ml-dtypes==0.3.2
more-itertools
mpmath==1.3.0
msgpack
multidict
multiprocess==0.70.16
murmurhash==1.0.10
namex==0.0.8
navigator-updater
nbclient
nbconvert
nbformat
nest-asyncio
networkx==3.3
nltk==3.5
notebook
notebook_shim
numexpr
numpy
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
overrides
packaging
pandas
pandocfilters
parso
partd
pathlib
pexpect
pillow
pkce
pkginfo
platformdirs
pluggy==1.6.0
ply==3.11
preprocessor==1.1.3
preshed==3.0.9
prometheus-client
prompt-toolkit
proto-plus==1.24.0
protobuf==4.25.3
psutil
psycopg2-binary==2.9.11
ptyprocess
pure-eval
py-cpuinfo
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycosat
pycparser
pycryptodomex==3.23.0
pydantic
pygame==2.6.1
Pygments
PyJWT
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-sip
PyQtWebEngine==5.15.6
PySocks
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil
python-docx==0.8.10
python-dotenv
python-json-logger
python-snappy
pytz
PyYAML
pyzmq
QtPy
referencing
regex==2024.9.11
requests
requests-oauthlib==2.0.0
requests-toolbelt
rfc3339-validator
rfc3986-validator
rich==13.7.1
rpds-py
rsa==4.9
ruamel-yaml-conda
ruamel.yaml
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
semver
Send2Trash
sentry-sdk==2.15.0
seqeval==1.2.2
setproctitle==1.3.3
shellingham==1.5.4
sip
six
smart-open==7.0.4
smmap==5.0.1
sniffio
sortedcontainers
soupsieve
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.4.8
stack-data
starlette==0.49.3
sympy==1.13.0
tables
tblib
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminado
tf_keras==2.17.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.11.0
tinycss2
tokenizers==0.20.0
toolz
torch==2.3.1
tornado
tqdm
traitlets
transformers==4.45.2
truststore
typer==0.12.3
typing_extensions==4.15.0
tzdata
ujson
unrar==0.4
uritemplate==4.1.1
urllib3
uvicorn==0.38.0
wandb==0.18.5
wasabi==1.1.3
wcwidth
weasel==0.4.1
webencodings==0.5.1
websocket-client
Werkzeug==3.0.3
widgetsnbextension
wrapt==1.16.0
xxhash==3.5.0
xyzservices
yarl
zict
zipp==2.2.0
zstandard
flask
flask-cors
psycopg2-binary
